Smoke Effect on Death
Trying to get there to be a smoke effect when the player dies. Got it with the code in particle controller, but it only works if the code is also in the enemy script but I want it to work as its own script so it can be used generally.